chore(state): move inline test modules to tests dir (closes #144)
